Widget Android

+++ iOS/G3MiOSSDK/Specific/Widget/G3MWidget_iOS.mm
@@ -134,7 +134,7 @@
                                 Planet::createEarth(),
                                 comp,
                                 width, height,
-                                Color::fromRGB(0, 0.1, 0.2, 1)); 
+                                Color::fromRGB((float)0, (float)0.1, (float)0.2, (float)1)); 
 
     glEnable(GL_DEPTH_TEST);
     glEnable(GL_CULL_FACE);
